+++ Libplanet.Explorer.Tests/GeneratedBlockChainFixture.cs
@@ -85,8 +85,7 @@ public class GeneratedBlockChainFixture
         var actionEvaluator = new ActionEvaluator(
             _ => policy.BlockAction,
             new BlockChainStates(store, stateStore),
-            TypedActionLoader.Create(typeof(SimpleAction).Assembly, typeof(SimpleAction)),
-            null);
+            TypedActionLoader.Create(typeof(SimpleAction).Assembly, typeof(SimpleAction)));
         Block genesisBlock = BlockChain.ProposeGenesisBlock(
             actionEvaluator,
             transactions: PrivateKeys