+++ Libplanet.Tests/Blockchain/BlockChainTest.cs
@@ -66,8 +66,7 @@ namespace Libplanet.Tests.Blockchain
                 new ActionEvaluator(
                     _ => _policy.BlockAction,
                     blockChainStates: new BlockChainStates(_fx.Store, _fx.StateStore),
-                    actionTypeLoader: new SingleActionLoader(typeof(DumbAction)),
-                    feeCalculator: null),
+                    actionTypeLoader: new SingleActionLoader(typeof(DumbAction))),
                 renderers: new[] { new LoggedActionRenderer(_renderer, Log.Logger) }
             );
             _renderer.ResetRecords();
@@ -141,8 +140,7 @@ namespace Libplanet.Tests.Blockchain
                 new ActionEvaluator(
                     _ => policy.BlockAction,
                     blockChainStates,
-                    new SingleActionLoader(typeof(DumbAction)),
-                    null));
+                    new SingleActionLoader(typeof(DumbAction))));
 
             Assert.Equal(chain1.Id, z.Id);
         }
@@ -188,8 +186,7 @@ namespace Libplanet.Tests.Blockchain
             var actionEvaluator = new ActionEvaluator(
                 _ => policy.BlockAction,
                 blockChainStates,
-                actionLoader,
-                null);
+                actionLoader);
             var nonce = 0;
             var txs = TestUtils.ValidatorSet.Validators
                 .Select(validator => Transaction.Create(
@@ -600,8 +597,7 @@ namespace Libplanet.Tests.Blockchain
                 var actionEvaluator = new ActionEvaluator(
                     _ => _policy.BlockAction,
                     new BlockChainStates(store, stateStore),
-                    new SingleActionLoader(typeof(DumbAction)),
-                    null);
+                    new SingleActionLoader(typeof(DumbAction)));
                 var genesis = ProposeGenesisBlock(
                     actionEvaluator,
                     ProposeGenesis(
@@ -632,8 +628,7 @@ namespace Libplanet.Tests.Blockchain
                     new ActionEvaluator(
                         _ => _policy.BlockAction,
                         blockChainStates: new BlockChainStates(store, stateStore),
-                        actionTypeLoader: new SingleActionLoader(typeof(DumbAction)),
-                        feeCalculator: null),
+                        actionTypeLoader: new SingleActionLoader(typeof(DumbAction))),
                     renderers: new[] { renderer }
                 );
 
@@ -1023,8 +1018,7 @@ namespace Libplanet.Tests.Blockchain
                 var actionEvaluator = new ActionEvaluator(
                     _ => _policy.BlockAction,
                     blockChainStates: new BlockChainStates(fx2.Store, fx2.StateStore),
-                    actionTypeLoader: new SingleActionLoader(typeof(DumbAction)),
-                    feeCalculator: null);
+                    actionTypeLoader: new SingleActionLoader(typeof(DumbAction)));
                 Block genesis2 = ProposeGenesisBlock(
                     actionEvaluator,
                     ProposeGenesis(
@@ -1081,8 +1075,7 @@ namespace Libplanet.Tests.Blockchain
             var actionEvaluator = new ActionEvaluator(
                 _ => policy.BlockAction,
                 new BlockChainStates(store, stateStore),
-                new SingleActionLoader(typeof(DumbAction)),
-                null);
+                new SingleActionLoader(typeof(DumbAction)));
             Block genesisWithTx = ProposeGenesisBlock(
                 actionEvaluator,
                 ProposeGenesis(
@@ -1125,8 +1118,7 @@ namespace Libplanet.Tests.Blockchain
                 new ActionEvaluator(
                     _ => policy.BlockAction,
                     blockChainStates,
-                    new SingleActionLoader(typeof(DumbAction)),
-                    null));
+                    new SingleActionLoader(typeof(DumbAction))));
 
             Block b = chain.Genesis;
             Address[] addresses = new Address[30];
@@ -1180,8 +1172,7 @@ namespace Libplanet.Tests.Blockchain
                 new ActionEvaluator(
                     _ => policy.BlockAction,
                     blockChainStates,
-                    new SingleActionLoader(typeof(DumbAction)),
-                    null));
+                    new SingleActionLoader(typeof(DumbAction))));
 
             Block b = chain.Genesis;
             for (int i = 0; i < 20; ++i)
@@ -1256,8 +1247,7 @@ namespace Libplanet.Tests.Blockchain
                 new ActionEvaluator(
                     _ => policy.BlockAction,
                     blockChainStates,
-                    new SingleActionLoader(typeof(DumbAction)),
-                    null));
+                    new SingleActionLoader(typeof(DumbAction))));
 
             Assert.All(chain.GetStates(addresses), Assert.Null);
             foreach (var address in addresses)
@@ -1329,8 +1319,7 @@ namespace Libplanet.Tests.Blockchain
                     new ActionEvaluator(
                         _ => _blockChain.Policy.BlockAction,
                         blockChainStates: new BlockChainStates(emptyFx.Store, emptyFx.StateStore),
-                        actionTypeLoader: new SingleActionLoader(typeof(DumbAction)),
-                        feeCalculator: null));
+                        actionTypeLoader: new SingleActionLoader(typeof(DumbAction))));
                 var fork = BlockChain.Create(
                     _blockChain.Policy,
                     new VolatileStagePolicy(),
@@ -1340,8 +1329,7 @@ namespace Libplanet.Tests.Blockchain
                     new ActionEvaluator(
                         _ => _blockChain.Policy.BlockAction,
                         blockChainStates: new BlockChainStates(forkFx.Store, forkFx.StateStore),
-                        actionTypeLoader: new SingleActionLoader(typeof(DumbAction)),
-                        feeCalculator: null));
+                        actionTypeLoader: new SingleActionLoader(typeof(DumbAction))));
                 fork.Append(b1, CreateBlockCommit(b1));
                 fork.Append(b2, CreateBlockCommit(b2));
                 Block b5 = fork.ProposeBlock(
@@ -1690,8 +1678,7 @@ namespace Libplanet.Tests.Blockchain
             var actionEvaluator = new ActionEvaluator(
                 _ => blockPolicy.BlockAction,
                 blockChainStates: chainStates,
-                actionTypeLoader: new SingleActionLoader(typeof(DumbAction)),
-                feeCalculator: null);
+                actionTypeLoader: new SingleActionLoader(typeof(DumbAction)));
             Block genesisBlock = ProposeGenesisBlock(
                 actionEvaluator,
                 ProposeGenesis(GenesisProposer.PublicKey),
@@ -1916,8 +1903,7 @@ namespace Libplanet.Tests.Blockchain
             var actionEvaluator = new ActionEvaluator(
                 _ => policy.BlockAction,
                 blockChainStates,
-                new SingleActionLoader(typeof(DumbAction)),
-                null);
+                new SingleActionLoader(typeof(DumbAction)));
             BlockChain blockChain = BlockChain.Create(
                 policy,
                 new VolatileStagePolicy(),
@@ -1957,8 +1943,7 @@ namespace Libplanet.Tests.Blockchain
             var actionEvaluator = new ActionEvaluator(
                 _ => policy.BlockAction,
                 blockChainStates,
-                new SingleActionLoader(typeof(DumbAction)),
-                null);
+                new SingleActionLoader(typeof(DumbAction)));
             var genesisBlockA = BlockChain.ProposeGenesisBlock(actionEvaluator);
             var genesisBlockB = BlockChain.ProposeGenesisBlock(actionEvaluator);
 
@@ -2037,8 +2022,7 @@ namespace Libplanet.Tests.Blockchain
             var actionEvaluator = new ActionEvaluator(
                 _ => policy.BlockAction,
                 new BlockChainStates(store, stateStore),
-                new SingleActionLoader(typeof(DumbAction)),
-                null);
+                new SingleActionLoader(typeof(DumbAction)));
             var genesisWithTx = ProposeGenesisBlock(
                 actionEvaluator,
                 ProposeGenesis(GenesisProposer.PublicKey, new[] { genesisTx }),
@@ -2101,8 +2085,7 @@ namespace Libplanet.Tests.Blockchain
             var actionEvaluator = new ActionEvaluator(
                 _ => policy.BlockAction,
                 new BlockChainStates(storeFixture.Store, storeFixture.StateStore),
-                new SingleActionLoader(typeof(SetValidator)),
-                null);
+                new SingleActionLoader(typeof(SetValidator)));
             Block genesis = BlockChain.ProposeGenesisBlock(
                 actionEvaluator,
                 privateKey: privateKey,