+++ Libplanet.Tests/TestUtils.cs
@@ -609,8 +609,7 @@ Actual (C# array lit):   new byte[{actual.LongLength}] {{ {actualRepr} }}";
             var actionEvaluator = new ActionEvaluator(
                     _ => policy.BlockAction,
                     blockChainStates: blockChainStates,
-                    actionTypeLoader: new SingleActionLoader(typeof(T)),
-                    feeCalculator: null);
+                    actionTypeLoader: new SingleActionLoader(typeof(T)));
 
             if (genesisBlock is null)
             {